requirements.txt
fastapi==0.103.1
colorama==0.4.6
python-dotenv==1.0.0